+++ blueprints/codes/eurocode/pr_en_1992_1_1_2023/chapter_8_ultimate_limit_states/formula_8_50.py
@@ -1,6 +1,6 @@
 """Formula 8.50 from prEN-1992-1-1:2023: Chapter 8 - Ultimate limit states (ULS)."""
 
-from blueprints.codes.eurocode.pr_en_1992_1_2023 import PR_EN_1992_1_1_2023
+from blueprints.codes.eurocode.pr_en_1992_1_1_2023 import PR_EN_1992_1_1_2023
 from blueprints.codes.formula import Formula
 from blueprints.codes.latex_formula import LatexFormula, latex_replace_symbols
 from blueprints.type_alias import DEG, N
