+++ blueprints/codes/eurocode/pr_en_1992_1_1_2023/chapter_8_ultimate_limit_states/formula_8_53.py
@@ -1,68 +1,65 @@
-"""Formula 8.52 from prEN 1992-1-1:2023: Chapter 8 - Ultimate Limit State."""
+"""Formula 8.53 from prEN 1992-1-1:2023: Chapter 8 - Ultimate Limit State."""
 
-from blueprints.codes.eurocode.pr_en_1992_1_2023 import PR_EN_1992_1_1_2023
+from blueprints.codes.eurocode.pr_en_1992_1_1_2023 import PR_EN_1992_1_1_2023
 from blueprints.codes.formula import Formula
 from blueprints.codes.latex_formula import LatexFormula, latex_replace_symbols
 from blueprints.type_alias import MM, NMM, N
 from blueprints.validations import raise_if_less_or_equal_to_zero, raise_if_negative
 
 
-class Form8Dot52CompressiveChordForceDueToShear(Formula):
-    r"""Class representing formula 8.52 for the calculation of the compressive chord force due to shear."""
+class Form8Dot53TensileChordLimitForIntermediateSupportOrConcentratedLoads(Formula):
+    r"""Class representing formula 8.53 for the calculation of the tensile chord limit in case of direct intermediate
+    support or in the region of concentrated loads.
+    """
 
-    label = "8.52"
+    label = "8.53"
     source_document = PR_EN_1992_1_1_2023
 
     def __init__(
         self,
-        m_ed: NMM,
+        m_ed_max: NMM,
         z: MM,
-        n_vd: N,
         n_ed: N,
     ) -> None:
-        r"""[$F_{cd}$] Calculation of the compressive chord force due to shear [$N$].
+        r"""[$F_{td,max}$] Calculation of the chord force [$N$] in case of direct intermediate support or
+        concentrated loads.
 
-        prEN 1992-1-1:2023 art 8.2.3 (8) - Formula (8.52)
+        prEN 1992-1-1:2023 art.8.5(8) - Formula (8.53)
 
         Parameters
         ----------
-        m_ed : NMM
-            [$M_{Ed}$] Design moment at the section [$Nmm$].
+        m_ed_max : NMM
+            [$M_{Ed,max}$] Maximum design moment along the member [$Nmm$].
         z : MM
-            [$z$] Lever arm [$mm$].
-        n_vd : N
-            [$N_{Vd}$] Additional tensile axial force due to shear [$N$].
+            [$z$] Internal lever arm [$mm$].
         n_ed : N
-            [$N_{Ed}$] Axial force in the section [$N$].
+            [$N_{Ed}$] Design value of normal force in the section [$N$].
         """
         super().__init__()
-        self.m_ed = m_ed
+        self.m_ed_max = m_ed_max
         self.z = z
-        self.n_vd = n_vd
         self.n_ed = n_ed
 
     @staticmethod
     def _evaluate(
-        m_ed: NMM,
+        m_ed_max: NMM,
         z: MM,
-        n_vd: N,
         n_ed: N,
     ) -> N:
         """Evaluates the formula, for more information see the __init__ method."""
         raise_if_less_or_equal_to_zero(z=z)
-        raise_if_negative(m_ed=m_ed, n_vd=n_vd, n_ed=n_ed)
+        raise_if_negative(m_ed_max=m_ed_max, n_ed=n_ed)
 
-        return m_ed / z - (n_vd + n_ed) / 2
+        return m_ed_max / z + n_ed / 2
 
     def latex(self, n: int = 3) -> LatexFormula:
-        """Returns LatexFormula object for formula 8.52."""
-        _equation: str = r"\frac{M_{Ed}}{z} - \frac{N_{Vd} + N_{Ed}}{2}"
+        """Returns LatexFormula object for formula 8.53."""
+        _equation: str = r"\frac{M_{Ed,max}}{z} + \frac{N_{Ed}}{2}"
         _numeric_equation: str = latex_replace_symbols(
             _equation,
             {
-                r"M_{Ed}": f"{self.m_ed:.{n}f}",
+                r"M_{Ed,max}": f"{self.m_ed_max:.{n}f}",
                 r"z": f"{self.z:.{n}f}",
-                r"N_{Vd}": f"{self.n_vd:.{n}f}",
                 r"N_{Ed}": f"{self.n_ed:.{n}f}",
             },
             False,
@@ -70,15 +67,14 @@ class Form8Dot52CompressiveChordForceDueToShear(Formula):
         _numeric_equation_with_units: str = latex_replace_symbols(
             _equation,
             {
-                r"M_{Ed}": rf"{self.m_ed:.{n}f} \ Nmm",
+                r"M_{Ed,max}": rf"{self.m_ed_max:.{n}f} \ Nmm",
                 r"z": rf"{self.z:.{n}f} \ mm",
-                r"N_{Vd}": rf"{self.n_vd:.{n}f} \ N",
                 r"N_{Ed}": rf"{self.n_ed:.{n}f} \ N",
             },
             True,
         )
         return LatexFormula(
-            return_symbol=r"F_{cd}",
+            return_symbol=r"F_{td,max}",
             result=f"{self:.{n}f}",
             equation=_equation,
             numeric_equation=_numeric_equation,
