+++ tests/codes/eurocode/pr_en_1992_1_1_2023/chapter_8_ultimate_limit_states/test_formula_8_50.py
@@ -4,7 +4,7 @@ import math
 
 import pytest
 
-from blueprints.codes.eurocode.pr_en_1992_1_2023.chapter_8_ultimate_limit_states.formula_8_50 import (
+from blueprints.codes.eurocode.pr_en_1992_1_1_2023.chapter_8_ultimate_limit_states.formula_8_50 import (
     Form8Dot50AdditionalTensileForceDueToShear,
 )
 from blueprints.validations import LessOrEqualToZeroError
