+++ tests/codes/eurocode/pr_en_1992_1_1_2023/chapter_8_ultimate_limit_states/test_formula_8_52.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from blueprints.codes.eurocode.pr_en_1992_1_2023.chapter_8_ultimate_limit_states.formula_8_52 import Form8Dot52CompressiveChordForceDueToShear
+from blueprints.codes.eurocode.pr_en_1992_1_1_2023.chapter_8_ultimate_limit_states.formula_8_52 import Form8Dot52CompressiveChordForceDueToShear
 from blueprints.validations import LessOrEqualToZeroError, NegativeValueError
